requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.40.0
python-docx==1.1.2
reportlab==4.2.2